Dia 4-5-6 haciendo el game

+++ src/entidades/Personaje.java
@@ -29,7 +29,7 @@ public abstract class Personaje {
 
 
     protected void initHitBox(float x, float y, float width, float height) {
-        this.hitBox = new Rectangle2D.Float(x, y, width, height);
+        this.hitBox = new Rectangle2D.Float(x, y, (int)width, (int)height);
     }
 
     /**
